cleaned up webApp code
