+++ RaspiSoftware/WebApp/static/javascript/controls.js
@@ -6,45 +6,41 @@ socket.on('connect', function(){
     console.log('Websocket connected!');
 });      
 
-socket.on('testing deployment socket', function(){
-    console.log('getting message from deployment')
-});
-
-var update_power_slider = function() {
+function update_power_slider() {
     /* Update slider value */
     let slider = document.getElementById("motorPower");
     let display = document.getElementById("displayPower");
     display.innerHTML = slider.value;
 }
 
-var update_speed_slider = function() {
+function update_speed_slider() {
     /* Update slider value */
     let slider = document.getElementById("encoderSpeed");
     let display = document.getElementById("displaySpeed");
     display.innerHTML = slider.value;
 }
 
-var send_motor_power = function() {
-    /*  send motor power to python via Websockets
-        called by changing slider value */
+function send_motor_power() {
+    /*  Send motor power to to app.py via socket.io,
+        called by changing power slider value */
     var slider = document.getElementById("motorPower");
     console.log('Sending motor power!');
     socket.emit('new motor power', slider.value);
     $("#outgoing").load(location.href + " #outgoing");
 }
 
-var send_encoder_speed = function() {
-    /*  send encoder speed to python via Websockets
-        called by changing slider value */
+function send_encoder_speed() {
+    /*  Send encoder speed to app.py via socket.io,
+        called by changing speed slider value */
     var slider = document.getElementById("encoderSpeed");
     console.log('Sending encoder speed!');
     socket.emit('new motor target', slider.value);
     $("#outgoing").load(location.href + " #outgoing");
 }
 
-var send_encoder_position = function() {
-    /*  send encoder speed to python via Websockets
-        called by changing slider value */
+function send_encoder_position() {
+    /*  Send encoder position to app.py via socket.io,
+        called by changing position slider value */
     var input = document.getElementById("encoderPosition");
     position = input.value*2887 //Converting meters to encoder steps
     console.log('Sending encoder position!');
@@ -53,9 +49,9 @@ var send_encoder_position = function() {
     $("#outgoing").load(location.href + " #outgoing");
 }
 
-var send_serial_command = function() {
-    /*  send encoder speed to python via Websockets
-        called by changing slider value */
+function send_serial_command(){
+    /*  Send serial command to app.py via socket.io,
+        called by submitted a command in the "Manual" tab in Robot Controls */
     var input = document.getElementById("serialSend");
     command = input.value
     console.log('command type' + typeof(command))
@@ -66,7 +62,8 @@ var send_serial_command = function() {
 }
 
 function toggle_motor(){
-    /*  send enable or disable command to flask */
+    // Enable/Disable Motor and update appearance of motorStatus buttons 
+    // There are multiple motorStatus buttons because there is one nested in each tab
     console.log('inside toggle motor')
     var elem = document.getElementById("motorStatus1");
     if (elem.value=="Enable Motor"){
@@ -76,7 +73,6 @@ function toggle_motor(){
         document.getElementById("motorStatus2").value = "Disable Motor";
         document.getElementById("motorStatus3").value = "Disable Motor";
         document.getElementById("motorStatus4").value = "Disable Motor";
-
     }
     else{
         socket.emit('disable motor');
@@ -84,8 +80,6 @@ function toggle_motor(){
         document.getElementById("motorStatus2").value = "Enable Motor";
         document.getElementById("motorStatus3").value = "Enable Motor";
         document.getElementById("motorStatus4").value = "Enable Motor";
-
-
     } 
     $("#outgoing").load(location.href + " #outgoing");
 }
@@ -101,7 +95,6 @@ $("#v-pills-speed-tab").on('show.bs.tab', function() {
     console.log('toggle speed from tab');
     socket.emit('new motor mode', "2");
 });
-
 
 // Toggle Position
 $("#v-pills-position-tab").on('show.bs.tab', function() {
