+++ RaspiSoftware/WebApp/static/javascript/visualization.js
@@ -1,133 +1,94 @@
 // Establish socketIO connection
 var socket = io.connect('http://' + document.domain + ':' + location.port);
-
-var clients = [];
-
-// socket.on('connect', function(client){
-//   console.log('Websocket connected on visualization!');
-//   // console.log(' %s sockets connected on visualiztion', io.engine.clientsCount);
-//   clients.push(client);
-
-//   client.on('disconnect', function() {
-//     console.log("disconnect: ", socket.id);
-//     clients.splice(clients.indexOf(client), 1);
-//   });
-// });
 socket.on('connect', function(){
-  console.log('Websocket connected on visualization!');
+    console.log('Websocket connected on table_test!');
 });      
 
+// Set to false once the first row has been made in a table
+var logsNoEntries = true;
+var dataNoEntries = true
 
-// Google charts set up
-google.charts.load('current', {'packages':['table']});
-// Set a callback to run when the Google Visualization API is loaded.
-google.charts.setOnLoadCallback(initTable);
-google.charts.setOnLoadCallback(drawTable);
+// Initializes Datatables
+var robotLogsTable;
+var aquatrollDataTable;
+$(document).ready(function() {
+    robotLogsTable = $('#robotLogsTable').DataTable({
+        "searching": false,
+        "lengthMenu": [[5, 10, 25, 50, -1], [5, 10, 25, 50, "All"]]
+    });
+    aquatrollDataTable = $('#aquatrollDataTable').DataTable({
+        "lengthMenu": [[5, 10, 25, 50, -1], [5, 10, 25, 50, "All"]]
+    });
+} );
 
-var noEntries = true;
+socket.on('update robot logs table', function(jsonData){
+    // console.log("recieved update ROBOT table");
+    logsNoEntries = updateTable(jsonData, robotLogsTable, logsNoEntries); //update table with json from the database
+});
 
-function initTable(){
-  /** Sets DataTable and table global variables */
-  data = new google.visualization.DataTable();
-  data.addColumn('string', 'Sensor Name');
-  data.addColumn('string', 'Value');
-  data.addColumn('string', 'Time Stamp');
+socket.on('update Aqua TROLL data table', function(jsonData){
+    // console.log("recieved update DATA table");
+    dataNoEntries = updateTable(jsonData, aquatrollDataTable, dataNoEntries);
+  });
 
-  table = new google.visualization.Table(document.getElementById('sqlTable'));
+function updateTable(jsonData, table, noEntries) {
+    /*
+    Updates table with new values. If a sensor doesn't exist, create new row and populate it.
+    var jsonData = ["testName", "testValue", "2000-01-01 00:00:01"]
+    Parameters:
+        jsonData - data to be put into the table
+        table - table to be updated
+        noEntries - boolean checking whether there are any existing rows in the table
+    Returns:
+        noEntries - if noEntries was set to false locally, returns boolean so that the global variable
+                    can be updated.
+    */ 
+    // console.log("RUNNING UPDATE TABLE");
+    var parsedData = JSON.parse(jsonData);
+    var id = parsedData[0];
+    // Sort data according to the first column in ascending order
+    // table.order( [[ 0, 'asc' ]] );
+    // Gets exisiting ids of sensors in the tables
+    var existing_ids = get_existing_ids(table); // Type: List of ids in table already
+    // console.log("existing ids: " + existing_ids);
+
+    if (noEntries == true){
+        // If the table is empty, create the first row
+        table.row.add(parsedData).draw(false);
+        console.log("creating new row from noEntries")
+        noEntries = false;
+        existing_ids = get_existing_ids(table);// Update existing id to new value
+    }
+
+    if (existing_ids.includes(id)){
+        // If the table already has a row of that sensor, update data and redrew
+        // console.log('updating old row');
+        i = existing_ids.indexOf(id);
+        table.row(i).data(parsedData).draw(false);
+    }
+    else {
+        // Create a new row
+        // console.log('creating new row');
+        table.row.add(parsedData).draw(false);
+    }
+    return noEntries
 };
+
+function get_existing_ids(table){
+    var existing_ids = [];
+    var i;
+    for (i = 0; i < table.data().count(); i++) {
+        cell_id = table.cell(i,0).data();
+        existing_ids.push(cell_id);
+    }
+    return existing_ids;
+}
 
 // Update progress bar
 socket.on('update progress bar', function(progressVal){
-  console.log("updating progress bar");
-  console.log("progressvaltype:" + typeof progressVal)
-  console.log('progress val:' + progressVal)
-  var progressBar = document.getElementById("deploymentProgress");  
-  document.getElementsByClassName('progress-bar').item(0).setAttribute('aria-valuenow',progressVal);
-  document.getElementsByClassName('progress-bar').item(0).setAttribute('style','width:'+Number(progressVal)+'%');
+    console.log("updating progress bar");
+    // console.log('progress val:' + progressVal);
+    var progressBar = document.getElementById("deploymentProgress");  
+    document.getElementsByClassName('progress-bar').item(0).setAttribute('aria-valuenow',progressVal);
+    document.getElementsByClassName('progress-bar').item(0).setAttribute('style','width:'+Number(progressVal)+'%');
 });
-
-socket.on('update table', function(jsonData){
-  /** Calls function that updates and redraws table.
-   */
-  console.log("recieved update table on visualization");
-  updateTable(jsonData); //update table with json from the database
-  drawTable();
-  data.sort([{column: 2}, {column: 1}]);
-});
-
-function updateTable(jsonData) {
-  /** Updates table with new values. If a sensor doesn't exist, create new row and populate it.
-   *   var jsonData = ["testName", "testValue", "2000-01-01 00:00:01"]
-  */
-  // console.log("RUNNING UPDATE TABLE");
-  // console.log("jsonData: " + jsonData);
-  var parsedData = JSON.parse(jsonData);
-  // console.log(parsedData);
-  // Id of the sensor
-  var id = parsedData[0];
-  // console.log("id: " + id);
-  var value = parsedData[1];
-  var timestamp = parsedData[2];
-
-  //sort data according to column 1 in ascending order
-  data.sort({column: 0});
-  // Gets exisiting ids of sensors in the tables
-  var existing_ids = data.getDistinctValues(0); // Type: Object array in ascending order
-  // console.log("existing ids: " + existing_ids);
-
-  if (noEntries == true){
-    data.addRow(parsedData);
-    // console.log("creating new row from noEntries")
-    noEntries = false;
-    existing_ids = data.getDistinctValues(0); // Update existing id to new value
-  }
-
-  // Check whether sensor already exists in the table.
-  // for (i = 0; i < existing_ids.length; i++) {
-  //   // If row with matching sensor name exists, updates existing row
-  //   if (id == existing_ids[i]){
-  //     console.log("i: " + existing_ids[i]);
-  //     console.log('updating old row');
-  //     data.setCell(i, 1, value);  // setCell(row, column)
-  //     data.setCell(i, 2, timestamp);
-  //   } else {
-  //     console.log('creating new row');
-  //     // Create a new row and populate
-  //     data.addRow(parsedData);
-  //   }
-  // }
-
-  if (existing_ids.includes(id)){
-    i = existing_ids.indexOf(id)
-    // console.log("i: " + existing_ids[i]);
-    // console.log('updating old row');
-    data.setCell(i, 1, value);  // setCell(row, column)
-    data.setCell(i, 2, timestamp);
-  }
-  else {
-    console.log('creating new row');
-    // Create a new row and populate
-    data.addRow(parsedData);
-  }
-
-};
-
-var cssClassNames = {
-  // headerRow: 'someclass',
-  // tableRow: 'someclass',
-  // oddTableRow: 'someclass',
-  // selectedTableRow: 'someclass',
-  // hoverTableRow: 'someclass',
-  // headerCell: 'someclass',
-  // tableCell: 'someclass',
-  // rowNumberCell: 'someclass'
-  headerRow: 'thead', 
-}
-
-function drawTable(){
-  table.draw(data, {cssClassNames, showRowNumber: true, width: '100%', height: '100%', allowHtml: true});
-  var className = 'google-visualization-table-table';
-  $('.'+className).removeClass(className);
-};
-
-
