+++ RaspiSoftware/WebApp/templates/layout.jinja2
@@ -20,6 +20,7 @@
 
         <title>PRS Cable Bot</title>
     </head>
+    
     <style>
         .slidecontainer {
           width: 100%;
